Add files via upload
create a folder on your desktop, drop these two files in. create a folder within that folder call "files" and put shortcuts of notepad++, putty, pycharm, and VirtualBox in there and rename the following
notepad++ to notepad
putty should just be putty
PyCharm should just be pycharm
VirtualBox to vbox

if you want to add your own tools, right click to edit with notepad++ and change the file names to whatever you put in the files folder. enjoy!